requirements/base.txt
Django>=5.0,<5.1
python-dotenv==1.0.1
argon2-cffi==23.1.0
whitenoise==6.7.0
django-autocomplete-light==3.11.0
crispy-tailwind==1.0.3
django-crispy-forms==2.3
django-htmx-autocomplete==0.8.4
